Add other langlib modules to test suite.
