+++ langlib/lang.stream/src/main/java/org/ballerinalang/langlib/stream/Publish.java
@@ -17,13 +17,9 @@
 */
 package org.ballerinalang.langlib.stream;
 
-import org.ballerinalang.bre.Context;
-import org.ballerinalang.bre.bvm.BlockingNativeCallableUnit;
 import org.ballerinalang.jvm.Strand;
 import org.ballerinalang.jvm.values.StreamValue;
 import org.ballerinalang.model.types.TypeKind;
-import org.ballerinalang.model.values.BStream;
-import org.ballerinalang.model.values.BValue;
 import org.ballerinalang.natives.annotations.Argument;
 import org.ballerinalang.natives.annotations.BallerinaFunction;
 
@@ -40,14 +36,7 @@ import org.ballerinalang.natives.annotations.BallerinaFunction;
                 @Argument(name = "data", type = TypeKind.ANY)
         },
         isPublic = true)
-public class Publish extends BlockingNativeCallableUnit {
-
-    @Override
-    public void execute(Context context) {
-        BStream stream = (BStream) context.getRefArgument(0);
-        BValue data = context.getRefArgument(1);
-        stream.publish(data);
-    }
+public class Publish {
 
     public static void publish(Strand strand, StreamValue streamValue, Object data) {
         streamValue.publish(strand, data);
