+++ langlib/lang.stream/src/main/java/org/ballerinalang/langlib/stream/Subscribe.java
@@ -17,14 +17,10 @@
 */
 package org.ballerinalang.langlib.stream;
 
-import org.ballerinalang.bre.Context;
-import org.ballerinalang.bre.bvm.BlockingNativeCallableUnit;
 import org.ballerinalang.jvm.Strand;
 import org.ballerinalang.jvm.values.FPValue;
 import org.ballerinalang.jvm.values.StreamValue;
 import org.ballerinalang.model.types.TypeKind;
-import org.ballerinalang.model.values.BFunctionPointer;
-import org.ballerinalang.model.values.BStream;
 import org.ballerinalang.natives.annotations.Argument;
 import org.ballerinalang.natives.annotations.BallerinaFunction;
 
@@ -40,14 +36,7 @@ import org.ballerinalang.natives.annotations.BallerinaFunction;
                 @Argument(name = "func", type = TypeKind.ANY)
         },
         isPublic = true)
-public class Subscribe extends BlockingNativeCallableUnit {
-
-    @Override
-    public void execute(Context context) {
-        BStream stream = (BStream) context.getRefArgument(0);
-        BFunctionPointer functionPointer = (BFunctionPointer) context.getRefArgument(1);
-        stream.subscribe(functionPointer);
-    }
+public class Subscribe {
 
     public static void subscribe(Strand strand, StreamValue streamValue, FPValue<Object[], Object> fpValue) {
         streamValue.subscribe(fpValue);
