+++ langlib/langlib-test/build.gradle
@@ -34,11 +34,36 @@ dependencies {
     testCompile project(path: ':ballerina-test-utils', configuration: 'shadow')
     testCompile 'org.testng:testng'
 
+    testRuntime project(':ballerina-lang:annotations')
+    testRuntime project(':ballerina-lang:array')
+    testRuntime project(':ballerina-lang:decimal')
+    testRuntime project(':ballerina-lang:error')
+    testRuntime project(':ballerina-lang:floatingpoint')
+    testRuntime project(':ballerina-lang:future')
+    testRuntime project(':ballerina-lang:integer')
+    testRuntime project(':ballerina-lang:map')
+    testRuntime project(':ballerina-lang:object')
+    testRuntime project(':ballerina-lang:stream')
     testRuntime project(':ballerina-lang:string')
+    testRuntime project(':ballerina-lang:table')
+    testRuntime project(':ballerina-lang:value')
+    testRuntime project(':ballerina-lang:xml')
     testRuntime 'org.slf4j:slf4j-jdk14'
 
     baloTestImplementation project(path: ':ballerina-lang:annotations', configuration: 'baloImplementation')
+    baloTestImplementation project(path: ':ballerina-lang:array', configuration: 'baloImplementation')
+    baloTestImplementation project(path: ':ballerina-lang:decimal', configuration: 'baloImplementation')
+    baloTestImplementation project(path: ':ballerina-lang:error', configuration: 'baloImplementation')
+    baloTestImplementation project(path: ':ballerina-lang:floatingpoint', configuration: 'baloImplementation')
+    baloTestImplementation project(path: ':ballerina-lang:future', configuration: 'baloImplementation')
+    baloTestImplementation project(path: ':ballerina-lang:integer', configuration: 'baloImplementation')
+    baloTestImplementation project(path: ':ballerina-lang:map', configuration: 'baloImplementation')
+    baloTestImplementation project(path: ':ballerina-lang:object', configuration: 'baloImplementation')
+    baloTestImplementation project(path: ':ballerina-lang:stream', configuration: 'baloImplementation')
     baloTestImplementation project(path: ':ballerina-lang:string', configuration: 'baloImplementation')
+    baloTestImplementation project(path: ':ballerina-lang:table', configuration: 'baloImplementation')
+    baloTestImplementation project(path: ':ballerina-lang:value', configuration: 'baloImplementation')
+    baloTestImplementation project(path: ':ballerina-lang:xml', configuration: 'baloImplementation')
     baloTestImplementation project(path: ':ballerina-utils', configuration: 'baloImplementation')
 }
