--- langlib/langlib-test/src/main/ballerina/Ballerina.toml
@@ -1,3 +1,0 @@
-[project]
-org-name = "testing"
-version = "0.0.0"
