+++ settings.gradle
@@ -6,9 +6,9 @@ include(':ballerina-lang:annotations')
 include(':ballerina-lang:array')
 include(':ballerina-lang:decimal')
 include(':ballerina-lang:error')
-include(':ballerina-lang:float')
+include(':ballerina-lang:floatingpoint')
 include(':ballerina-lang:future')
-include(':ballerina-lang:int')
+include(':ballerina-lang:integer')
 include(':ballerina-lang:map')
 include(':ballerina-lang:object')
 include(':ballerina-lang:stream')
@@ -130,9 +130,9 @@ project(':ballerina-lang:annotations').projectDir = file('langlib/lang.annotatio
 project(':ballerina-lang:array').projectDir = file('langlib/lang.array')
 project(':ballerina-lang:decimal').projectDir = file('langlib/lang.decimal')
 project(':ballerina-lang:error').projectDir = file('langlib/lang.error')
-project(':ballerina-lang:float').projectDir = file('langlib/lang.float')
+project(':ballerina-lang:floatingpoint').projectDir = file('langlib/lang.float')
 project(':ballerina-lang:future').projectDir = file('langlib/lang.future')
-project(':ballerina-lang:int').projectDir = file('langlib/lang.int')
+project(':ballerina-lang:integer').projectDir = file('langlib/lang.int')
 project(':ballerina-lang:map').projectDir = file('langlib/lang.map')
 project(':ballerina-lang:object').projectDir = file('langlib/lang.object')
 project(':ballerina-lang:stream').projectDir = file('langlib/lang.stream')
